Change attachment_association to on_save for course syllabus
refs GROW-86
flag=disable_file_verifiers_in_public_syllabus
[skip-crystalball]

Test plan
- Create a new course with attachments linked
  in the syllabus body and a public syllabus
- Update an old course with attachments linked
  in the syllabus body (ensure the syllabus
  is public)
- Copy a course with attachments linked in the
  syllabus body
- Ensure non-logged in users can see the attachments
  in each case

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/381335
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Andras Kemeny <akemeny@instructure.com>
Product-Review: Andras Kemeny <akemeny@instructure.com>

+++ app/models/attachment_association.rb
@@ -67,6 +67,9 @@ class AttachmentAssociation < ActiveRecord::Base
     to_delete = currently_has - global_ids
     to_create = global_ids - currently_has
 
+    return if (to_create + to_delete).none?
+    raise "User is required to update attachment links" if user.blank? && !blank_user
+
     if to_delete.any?
       context.attachment_associations.where(context_concern:, attachment_id: to_delete).in_batches(of: 1000).destroy_all
     end
